refactor: move leave colony modal to own component, fix display names of home partials
