+++ src/components/v5/frame/ColonyHome/ColonyHome.tsx
@@ -1,14 +1,8 @@
 import React from 'react';
-import { defineMessages } from 'react-intl';
 
 import ColonyActionsTable from '~common/ColonyActionsTable';
-import { useColonyDashboardContext } from '~context/ColonyDashboardContext';
 import { useSetPageBreadcrumbs } from '~context/PageHeadingContext/hooks';
-import {
-  useColonySubscription,
-  useGetSelectedDomainFilter,
-  useMobile,
-} from '~hooks';
+import { useGetSelectedDomainFilter, useMobile } from '~hooks';
 import { useCreateTeamBreadcrumbs } from '~hooks/useTeamsBreadcrumbs';
 import {
   // @BETA: Disabled for now
@@ -19,10 +13,10 @@ import {
 import { formatText } from '~utils/intl';
 import { setQueryParamOnUrl } from '~utils/urls';
 import Link from '~v5/shared/Link';
-import Modal from '~v5/shared/Modal';
 
 import Agreements from './partials/Agreements';
 import DashboardHeader from './partials/DashboardHeader';
+import LeaveColonyModal from './partials/LeaveColonyModal';
 import Members from './partials/Members';
 import Objective from './partials/Objective';
 import ReputationChart from './partials/ReputationChart';
@@ -31,29 +25,10 @@ import TotalActions from './partials/TotalActions';
 
 const displayName = 'v5.frame.ColonyHome';
 
-const MSG = defineMessages({
-  leaveConfimModalTitle: {
-    id: `${displayName}.leaveConfimModalTitle`,
-    defaultMessage: 'Are you sure you want to leave this Colony?',
-  },
-  leaveConfirmModalSubtitle: {
-    id: `${displayName}.leaveConfirmModalSubtitle`,
-    defaultMessage: `Leaving this Colony will mean you are no longer able to access the Colony during the beta period. Ensure you have a copy of the invite code available to be able to re-join again.`,
-  },
-  leaveConfirmModalConfirmButton: {
-    id: `${displayName}.leaveConfirmModalConfirmButton`,
-    defaultMessage: 'Yes, leave this Colony',
-  },
-});
-
 const ColonyHome = () => {
   const isMobile = useMobile();
   const selectedDomain = useGetSelectedDomainFilter();
   const teamsBreadcrumbs = useCreateTeamBreadcrumbs();
-
-  const { isLeaveColonyModalOpen, closeLeaveColonyModal } =
-    useColonyDashboardContext();
-  const { handleUnwatch } = useColonySubscription();
 
   useSetPageBreadcrumbs(teamsBreadcrumbs);
 
@@ -106,22 +81,7 @@ const ColonyHome = () => {
           />
         </div>
       </div>
-      <Modal
-        title={formatText(MSG.leaveConfimModalTitle)}
-        subTitle={formatText(MSG.leaveConfirmModalSubtitle)}
-        isOpen={isLeaveColonyModalOpen}
-        onClose={() => closeLeaveColonyModal()}
-        onConfirm={() => {
-          closeLeaveColonyModal();
-          handleUnwatch();
-        }}
-        icon="warning-circle"
-        buttonMode="primarySolid"
-        confirmMessage={formatText(MSG.leaveConfirmModalConfirmButton)}
-        closeMessage={formatText({
-          id: 'button.cancel',
-        })}
-      />
+      <LeaveColonyModal />
     </div>
   );
 };
