+++ src/components/v5/frame/ColonyHome/partials/LeaveColonyModal/index.ts
@@ -1,0 +1,1 @@
+export { default } from './LeaveColonyModal';
